+++ SimulationCode/src/BarnesHutStepCuda.cpp
@@ -1,17 +1,17 @@
 #include "BarnesHutStepCuda.h"
 
-BarnesHutStepCuda::BarnesHutStepCuda(std::vector<double>& masses, unsigned numberOfBodies) {
+BarnesHutStepCuda::BarnesHutStepCuda(std::vector<float>& masses, unsigned numberOfBodies) {
     this->numberOfBodies = numberOfBodies;
     weights.resize(numberOfBodies);
     forces.resize(3*numberOfBodies);
 
-    //std::array<double, 6> boundariesForRoot = {-sizeFrame, sizeFrame, -sizeFrame, sizeFrame, -sizeFrame, sizeFrame};
+    //std::array<float, 6> boundariesForRoot = {-sizeFrame, sizeFrame, -sizeFrame, sizeFrame, -sizeFrame, sizeFrame};
     //root = new NodeBH(boundariesForRoot);
 
-    weights = std::vector<double>(masses.begin(), masses.end());
+    weights = std::vector<float>(masses.begin(), masses.end());
 
     randomGenerator = new RandomGenerators();
-    randomGenerator->initializeVelocities<std::vector<double>>(velocities, numberOfBodies);
+    randomGenerator->initializeVelocities<std::vector<float>>(velocities, numberOfBodies);
 
     c = new ComputationsBarnesHut(velocities, weights);
 }
@@ -21,6 +21,6 @@ BarnesHutStepCuda::~BarnesHutStepCuda() {
     delete randomGenerator;
 }
 
-void BarnesHutStepCuda::compute(tf3 &positions, double dt) {
+void BarnesHutStepCuda::compute(tf3 &positions, float dt) {
   c->BarnesHutBridge(positions, numberOfBodies, dt);
 }
