+++ SimulationCode/src/StepNaiveCUDA.cpp
@@ -1,13 +1,13 @@
 #include "StepNaiveCUDA.h"
 
-StepNaiveCuda::StepNaiveCuda(std::vector<double> masses, unsigned numberOfBodies) {
+StepNaiveCuda::StepNaiveCuda(std::vector<float> masses, unsigned numberOfBodies) {
   this->numberOfBodies = numberOfBodies;
   weights.resize(numberOfBodies);
   for (unsigned i = 0; i < numberOfBodies; i++) {
     weights[i] = masses[i];
   }
   randomGenerator = new RandomGenerators();
-  randomGenerator->initializeVelocities<thrust::host_vector<double>>(velocities, numberOfBodies);
+  randomGenerator->initializeVelocities<thrust::host_vector<float>>(velocities, numberOfBodies);
   c = new Computations(velocities, weights);
 }
 
@@ -18,7 +18,7 @@ StepNaiveCuda::~StepNaiveCuda() {
   delete randomGenerator;
 }
 
-void StepNaiveCuda::compute(tf3 &positions, double dt) {
+void StepNaiveCuda::compute(tf3 &positions, float dt) {
   c->NaiveSimBridgeThrust(positions, numberOfBodies, dt);
 }
 
@@ -30,4 +30,4 @@ void StepNaiveCuda::compute(tf3 &positions, double dt) {
 // https://groups.google.com/forum/#!topic/thrust-users/4EaWLGeJOO8
 // https://github.com/thrust/thrust/blob/master/examples/cuda/unwrap_pointer.cu
 // thrust::copy(weightsD.begin(), weightsD.end(),
-// std::ostream_iterator<double>(std::cout, " "));
+// std::ostream_iterator<float>(std::cout, " "));
